codet5-base.jfreechart.mid-6605.idx-49929.3.mutant

+++ jfreechart/src/main/java/org/jfree/data/category/DefaultCategoryDataset.java
@@ -89,7 +89,7 @@ public class DefaultCategoryDataset<R extends Comparable<R>, C extends Comparabl
      */
     @Override
     public int getRowCount() {
-        return this.data.getRowCount();
+    return (int) this.data.getRowCount(); // casting the double value to int
     }
 
     /**
